+++ src/components/StandardPlot/StandardPlotContainer.js
@@ -1,10 +1,8 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
 import PropTypes from "prop-types";
-import {
-  getStandardPlotDataById,
-  getSelectedFieldObjectById,
-} from "../../states/standardplots/selector";
+import { getSelectedFieldObjectById } from "../../states/standardplots/utils";
+import { getStandardPlotDataById } from "../../states/standardplots/selector";
 import {
   getSpecById,
   getFacetById,
